requirements.txt
python-dotenv
cvat-sdk